requirements.txt
Flask==3.0.0
APScheduler==3.10.4
requests==2.32.4
beautifulsoup4==4.12.2
python-dotenv==1.0.0
Werkzeug==3.1.5
jsonpath-ng>=1.6.0
openai>=1.0.0
cryptography>=41.0.0

# Optional: proxy support (SOCKS). Plan: requests[socks]
requests[socks]==2.32.4

# Testing (plan: unit + integration + mocked AI)
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: screenshot-on-match (Phase 3.2). Then run: playwright install chromium
playwright>=1.40.0
